[abstract] Added abstract models, extending old base ones
See https://github.com/bennylope/django-organizations/issues/118
